Updated auth system to be stateless.
Updated the auth on the front end to maintain a key and token for api
authentication via cookies so that the api may remain stateless. Updated
the API to return properly formatted users with the new api key and
token. Also added the is authenticated functionality to revalidate the
passed key and tokens on page refresh.