Changed all api methods to POST
Currently the API server is 404ing every time I try to set the
authentication header. In the name of finishing this part of the project
I am creating a workaround where I pass the key and token pair to the
server as post parameters for authentication. But because I intend to
implement the appropriate authentication headers later I want to still
run the functions through "GET" and "DELETE" method names.