Added API Auth service to keep authentication centralized.
Removed API middleware auth and made it route middleware instead so that
I could define roles for routes or groups of routes - thus allowing me
to have 'public' or unauthenticated routes as well as 'member' vs
'admin' routes. Permissions may later be moved to the database for what
users may access which routes.