Bug Fix: Infinite loop was occuring when accessing login and signup.
Login and signup check if the user is logged in and redirect to the
dashboard. Since init was returning true it thought it the user was
logged in but when trying to access the member page the user wasnt
logged in so they were redirected back to login (hence loop of doom).