Email Service Conversion
I broke up the api.mailer class into several files. EmailService now
handles the sending of emails using the PhpMailer library.  The
RouteEmailsController abstract class provides the email templating
methods to create email bodies. And the classes extending the
RouteEmailsController provide the variables for the templates to replace
within the templates and trigger email events to occur on Routing.